refactor(memory): extract recall-tool pipeline into sibling modules (#445 PR-A)
Split recall-tool.ts god functions into focused modules to clear slop
CRITICAL findings while preserving MCP behavior and existing test coverage.
